+++ src/Algebra/Geometric/Cl3.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE Safe #-}
 {-# LANGUAGE GADTSyntax #-}
 {-# LANGUAGE ViewPatterns #-}
 {-# LANGUAGE DeriveDataTypeable #-}
@@ -8,15 +7,23 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE MultiWayIf #-}
 
+
+#ifndef O_LIQUID
+-- Turn off Safe Haskell language extension due to liquid-base re-exports
+{-# LANGUAGE Safe #-}
+#endif
+
 #if __GLASGOW_HASKELL__ == 810
 -- Work around to fix GHC Issue #15304, issue popped up again in GHC 8.10, it should be fixed in GHC 8.12
 -- This code is meant to reproduce MR 2608 for GHC 8.10
 {-# OPTIONS_GHC -funfolding-keeness-factor=1 -funfolding-use-threshold=80 #-}
 #endif
 
+
+
 --------------------------------------------------------------------------------------------
 -- |
--- Copyright   :  (C) 2017-2020 Nathan Waivio
+-- Copyright   :  (C) 2017-2022 Nathan Waivio
 -- License     :  BSD3
 -- Maintainer  :  Nathan Waivio <nathan.waivio@gmail.com>
 -- Stability   :  Stable
@@ -43,6 +50,20 @@ module Algebra.Geometric.Cl3
  showOctave,
  -- * Eliminate grades that are less than 'tol' to use a simpler Constructor
  reduce, tol,
+#ifndef O_NO_STORABLE
+ -- * Compact Storable types for the Cl3 Constructors with smart constructors
+ Cl3_R, toCl3_R, fromCl3_R,
+ Cl3_V3, toCl3_V3, fromCl3_V3,
+ Cl3_BV, toCl3_BV, fromCl3_BV,
+ Cl3_I, toCl3_I, fromCl3_I,
+ Cl3_PV, toCl3_PV, fromCl3_PV,
+ Cl3_H, toCl3_H, fromCl3_H,
+ Cl3_C, toCl3_C, fromCl3_C,
+ Cl3_BPV, toCl3_BPV, fromCl3_BPV,
+ Cl3_ODD, toCl3_ODD, fromCl3_ODD,
+ Cl3_TPV, toCl3_TPV, fromCl3_TPV,
+ Cl3_APS, toCl3_APS, fromCl3_APS,
+#endif
 #ifndef O_NO_RANDOM
  -- * Random Instances
  randR, rangeR,
@@ -64,19 +85,22 @@ module Algebra.Geometric.Cl3
  -- * Helpful Functions
  eigvals, hasNilpotent,
  spectraldcmp, project,
- mIx, timesI
+ mIx, timesI,
+ abssignum
 ) where
 
 #ifndef O_NO_DERIVED
 import Data.Data (Typeable, Data)
 import GHC.Generics (Generic)
+import Text.Read (Read,readPrec)
 #endif
 
-
 import Control.DeepSeq (NFData,rnf)
+
+#ifndef O_NO_STORABLE
 import Foreign.Storable (Storable, sizeOf, alignment, peek, poke)
 import Foreign.Ptr (Ptr, plusPtr, castPtr)
-
+#endif
 
 #ifndef O_NO_RANDOM
 import System.Random (RandomGen, Random, randomR, random)
@@ -136,6 +160,7 @@ instance Show Cl3 where
 #endif
 
 
+-- | Cl3 can be reduced to a normal form.
 instance NFData Cl3 where
   rnf !_ = ()
 
@@ -955,90 +980,20 @@ instance Num Cl3 where
                                                                                 (a0*b123 + a123*b0 + a1*b23 + a23*b1 + a2*b31 + a31*b2 + a3*b12 + a12*b3)
 
 
+
+
   -- |'abs' is the spectral norm aka the spectral radius
   -- it is the largest singular value. This function may need to be fiddled with
   -- to make the math a bit safer wrt overflows.  This makes use of the largest
   -- singular value, if the littlest singular value is zero then the element is not
   -- invertable, we can see here that R, C, V3, BV, and H are all invertable, and
   -- by implication R, C, and H are division algebras.
-  abs (R a0) = R (abs a0) -- absolute value of a real number
-  abs (V3 a1 a2 a3) = R (sqrt (a1^2 + a2^2 + a3^2)) -- magnitude of a vector
-  abs (BV a23 a31 a12) = R (sqrt (a23^2 + a31^2 + a12^2)) -- magnitude of a bivector
-  abs (I a123) = R (abs a123) -- magnitude of a Imaginary number
-  abs (PV a0 a1 a2 a3) = R (reimMag a0 a1 a2 a3)
-  abs (TPV a23 a31 a12 a123) = R (reimMag a123 a23 a31 a12)
-  abs (H a0 a23 a31 a12) = R (sqrt (a0^2 + a23^2 + a31^2 + a12^2)) -- largest singular value
-  abs (C a0 a123) = R (sqrt (a0^2 + a123^2)) -- magnitude of a complex number
-  abs (BPV a1 a2 a3 a23 a31 a12) =
-    let x = sqrt ((a1*a31 - a2*a23)^2 + (a1*a12 - a3*a23)^2 + (a2*a12 - a3*a31)^2) -- core was duplicating this computation added let to hopefully reduce the duplication
-    in R (sqrt (a1^2 + a23^2 + a2^2 + a31^2 + a3^2 + a12^2 + x + x))
-  abs (ODD a1 a2 a3 a123) = R (sqrt (a1^2 + a2^2 + a3^2 + a123^2))
-  abs (APS a0 a1 a2 a3 a23 a31 a12 a123) =
-    let x = sqrt ((a0*a1 + a123*a23)^2 + (a0*a2 + a123*a31)^2 + (a0*a3 + a123*a12)^2 +
-                  (a2*a12 - a3*a31)^2 + (a3*a23 - a1*a12)^2 + (a1*a31 - a2*a23)^2) -- core was duplicating this computation added let to hopefully reduce the duplication
-    in R (sqrt (a0^2 + a1^2 + a2^2 + a3^2 + a23^2 + a31^2 + a12^2 + a123^2 + x + x))
+  abs cl3 = fst $ abssignum cl3
 
 
   -- |'signum' satisfies the Law "abs x * signum x == x"
   -- kind of cool: signum of a vector is it's unit vector.
-  signum (R a0) = R (signum a0)
-  signum (V3 a1 a2 a3) =
-    let mag = sqrt (a1^2 + a2^2 + a3^2)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else V3 (invMag * a1) (invMag * a2) (invMag * a3)
-  signum (BV a23 a31 a12) =
-    let mag = sqrt (a23^2 + a31^2 + a12^2)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else BV (invMag * a23) (invMag * a31) (invMag * a12)
-  signum (I a123) = I (signum a123)
-  signum (PV a0 a1 a2 a3) =
-    let mag = reimMag a0 a1 a2 a3
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else PV (invMag * a0) (invMag * a1) (invMag * a2) (invMag * a3)
-  signum (H a0 a23 a31 a12) =
-    let mag = sqrt (a0^2 + a23^2 + a31^2 + a12^2)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else H (invMag * a0) (invMag * a23) (invMag * a31) (invMag * a12)
-  signum (C a0 a123) =
-    let mag = sqrt (a0^2 + a123^2)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else C (invMag * a0) (invMag * a123)
-  signum (BPV a1 a2 a3 a23 a31 a12) =
-    let x = sqrt ((a1*a31 - a2*a23)^2 + (a1*a12 - a3*a23)^2 + (a2*a12 - a3*a31)^2)
-        mag = sqrt (a1^2 + a23^2 + a2^2 + a31^2 + a3^2 + a12^2 + x + x)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else BPV (invMag * a1) (invMag * a2) (invMag * a3) (invMag * a23) (invMag * a31) (invMag * a12)
-  signum (ODD a1 a2 a3 a123) =
-    let mag = sqrt (a1^2 + a2^2 + a3^2 + a123^2)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else ODD (invMag * a1) (invMag * a2) (invMag * a3) (invMag * a123)
-  signum (TPV a23 a31 a12 a123) =
-    let mag = reimMag a123 a23 a31 a12
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else TPV (invMag * a23) (invMag * a31) (invMag * a12) (invMag * a123)
-  signum (APS a0 a1 a2 a3 a23 a31 a12 a123) =
-    let x = sqrt ((a0*a1 + a123*a23)^2 + (a0*a2 + a123*a31)^2 + (a0*a3 + a123*a12)^2 + (a2*a12 - a3*a31)^2 + (a3*a23 - a1*a12)^2 + (a1*a31 - a2*a23)^2)
-        mag = sqrt (a0^2 + a1^2 + a2^2 + a3^2 + a23^2 + a31^2 + a12^2 + a123^2 + x + x)
-        invMag = recip mag
-    in if mag == 0
-       then R 0
-       else APS (invMag * a0) (invMag * a1) (invMag * a2) (invMag * a3) (invMag * a23) (invMag * a31) (invMag * a12) (invMag * a123)
+  signum cl3 = snd $ abssignum cl3
 
 
   -- |'fromInteger'
@@ -1072,7 +1027,7 @@ reimMag :: Double -> Double -> Double -> Double -> Double
 reimMag v0 v1 v2 v3 =
   let sumsqs = v1^2 + v2^2 + v3^2
       x = abs v0 * sqrt sumsqs
-  in sqrt (v0^2 + sumsqs + x + x)
+  in sqrt (v0^2 + sumsqs + 2*x)
 
 -- |Cl(3,0) has a Fractional instance
 instance Fractional Cl3 where
@@ -1111,6 +1066,8 @@ instance Floating Cl3 where
     in C (expa0 * cos a123) (expa0 * sin a123)
   exp cliffor = spectraldcmp exp exp' cliffor
 
+
+
   --
   log (R a0)
     | a0 >= 0 = R (log a0)
@@ -1122,6 +1079,7 @@ instance Floating Cl3 where
     | otherwise = C (log.abs $ a123) (signum a123 * (pi/2))
   log (C a0 a123) = C (log (a0^2 + a123^2) / 2) (atan2 a123 a0)
   log cliffor = spectraldcmp log log' cliffor
+
 
   --
   sqrt (R a0)
@@ -1158,7 +1116,7 @@ instance Floating Cl3 where
   tan (I a123)
     | a123 == 0 = R 0
     | otherwise = I (tanh a123)
-  tan (C a0 a123) =
+  tan (reduce -> C a0 a123) =
     let
       m = x2^2 + y2^2
       x1 = sinx*coshy
@@ -1171,6 +1129,8 @@ instance Floating Cl3 where
       coshy = cosh a123
     in C ((x1*x2 + y1*y2)/m) ((x2*y1 - x1*y2)/m)
   tan cliffor = spectraldcmp tan tan' cliffor
+
+
 
   --
   asin (R a0)
@@ -1340,7 +1300,7 @@ instance Floating Cl3 where
   --
   tanh (R a0) = R (tanh a0)
   tanh (I a123) = I (tan a123)
-  tanh (C a0 a123) =
+  tanh (reduce -> C a0 a123) =
     let
       m = x2^2 + y2^2
       x1 = cosy*sinhx
@@ -1963,6 +1923,49 @@ timesI (ODD a1 a2 a3 a123) = H (negate a123) a1 a2 a3
 timesI (TPV a23 a31 a12 a123) = PV (negate a123) (negate a23) (negate a31) (negate a12)
 timesI (APS a0 a1 a2 a3 a23 a31 a12 a123) = APS (negate a123) (negate a23) (negate a31) (negate a12) a1 a2 a3 a0
 
+-- | 'abssignum' is a more effecient '\cl3 -> (abs cl3, signum cl3)'
+-- So 'abs' is always R and 'signum' is the same type of constructor as the input
+-- 'signum' is the element divided by its largest singular value 'abs'
+abssignum :: Cl3 -> (Cl3,Cl3)
+abssignum cl3 =
+  let (R m0) = absolute cl3
+  in if m0 == 0
+     then (R 0, R 0) -- (abs 0 == 0, signum 0 == 0)
+     else (R m0, cl3/(R m0))
+
+absolute :: Cl3 -> Cl3
+absolute (R a0) = R (abs a0)
+absolute (V3 a1 a2 a3) = let m = rss3 a1 a2 a3 in R m
+absolute (BV a23 a31 a12) = let m = rss3 a23 a31 a12 in R m
+absolute (I a123) = R (abs a123)
+absolute (PV a0 a1 a2 a3) = let m = reimMag a0 a1 a2 a3 in R m
+absolute (H a0 a23 a31 a12) = let m = rss4 a0 a23 a31 a12 in R m
+absolute (C a0 a123) = let m = rss2 a0 a123 in R m
+absolute (BPV a1 a2 a3 a23 a31 a12) = let mag0 = rss3 (a1*a31 - a2*a23) (a1*a12 - a3*a23) (a2*a12 - a3*a31)
+                                          m = sqrt $ a1^2 + a23^2 + a2^2 + a31^2 + a3^2 + a12^2 + 2*mag0 in R m
+absolute (ODD a1 a2 a3 a123) = let m = rss4 a1 a2 a3 a123 in R m
+absolute (TPV a23 a31 a12 a123) = let m = reimMag a123 a23 a31 a12 in R m
+absolute (APS a0 a1 a2 a3 a23 a31 a12 a123) = let mag0 = sqrt $ (a0*a1 + a123*a23)^2 + (a0*a2 + a123*a31)^2 + (a0*a3 + a123*a12)^2 + (a2*a12 - a3*a31)^2 + (a3*a23 - a1*a12)^2 + (a1*a31 - a2*a23)^2
+                                                  m = sqrt $ a0^2 + a1^2 + a2^2 + a3^2 + a23^2 + a31^2 + a12^2 + a123^2 + 2*mag0 in R m
+
+rss2 :: Double -> Double -> Double
+rss2 a0 a123 = sqrt $ a0^2 + a123^2
+
+rss3 :: Double -> Double -> Double -> Double
+rss3 x y z = sqrt $ x^2 + y^2 + z^2
+
+rss4 :: Double -> Double -> Double -> Double -> Double
+rss4 t x y z = sqrt $ t^2 + x^2 + y^2 + z^2
+
+
+
+#ifdef O_LIQUID
+tol :: Cl3
+tol = R 0
+
+tol' :: Double
+tol' = 0
+#else
 -- | 'tol' currently 128*eps
 tol :: Cl3
 {-# INLINE tol #-}
@@ -1971,7 +1974,7 @@ tol = R 1.4210854715202004e-14
 tol' :: Double
 {-# INLINE tol' #-}
 tol' = 1.4210854715202004e-14
-
+#endif
 
 -- | 'bar' is a Clifford Conjugate, the vector grades are negated
 bar :: Cl3 -> Cl3
@@ -2208,6 +2211,7 @@ atanh' :: Cl3 -> Cl3
 atanh' = recip.(1-).(^2)  -- pole at +/-1
 
 
+#ifndef O_NO_STORABLE
 -------------------------------------------------------------------
 -- 
 -- Instance of Cl3 types with the "Foreign.Storable" library.
@@ -2219,6 +2223,9 @@ atanh' = recip.(1-).(^2)  -- pole at +/-1
 
 -- | Cl3 instance of Storable uses the APS constructor as its standard interface.
 -- "peek" returns a cliffor constructed with APS. "poke" converts a cliffor to APS.
+-- For a more compact storing of constructors other than APS use the storable
+-- subtypes Cl3_R, Cl3_V3, Cl3_BV, Cl3_I, Cl3_PV, Cl3_H, Cl3_C, Cl3_BPV,
+-- Cl3_ODD, Cl3_TPV.
 instance Storable Cl3 where
   sizeOf _ = 8 * sizeOf (undefined :: Double)
   alignment _ = sizeOf (undefined :: Double)
@@ -2248,6 +2255,462 @@ instance Storable Cl3 where
         offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
   poke _ _ = error "Serious Issues with poke in Cl3.Storable"
 
+
+-- | 'Cl3_R' a compact storable data type for R.
+data Cl3_R where
+  Cl3_R :: !Double -> Cl3_R
+
+-- | 'toCl3_R' converts a Cl3 value constructed with R to its compact form.
+toCl3_R :: Cl3 -> Cl3_R
+toCl3_R (R a0) = Cl3_R a0
+toCl3_R err = error $ "Please don't try and cast something that's not R to Cl3_R, Got: " ++ show err
+
+-- | 'fromCl3_R' converts the compact Cl3_R type back to a Cl3 type.
+fromCl3_R :: Cl3_R -> Cl3
+fromCl3_R (Cl3_R a0) = R a0
+
+instance Show Cl3_R where
+  show = show.fromCl3_R
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_R where
+  readPrec = toCl3_R <$> readPrec
+#endif
+
+instance Storable Cl3_R where
+  sizeOf _ = sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a0 <- peek offset
+    return $ Cl3_R a0
+      where
+        offset = (castPtr ptr :: Ptr Double)
+
+  poke ptr (Cl3_R a0) = do
+    poke offset a0
+      where
+        offset = (castPtr ptr :: Ptr Double)
+
+
+-- | 'Cl3_V3' a compact storable data type for V3.
+data Cl3_V3 where
+  Cl3_V3 :: !Double -> !Double -> !Double -> Cl3_V3
+
+-- | 'toCl3_V3' converts a Cl3 value constructed with V3 to its compact form.
+toCl3_V3 :: Cl3 -> Cl3_V3
+toCl3_V3 (V3 a1 a2 a3) = Cl3_V3 a1 a2 a3
+toCl3_V3 err = error $ "Please don't try and cast something that's not V3 to Cl3_V3, Got: " ++ show err
+
+-- | 'fromCl3_V3' converts the compact Cl3_V3 type back to a Cl3 type.
+fromCl3_V3 :: Cl3_V3 -> Cl3
+fromCl3_V3 (Cl3_V3 a1 a2 a3) = V3 a1 a2 a3
+
+instance Show Cl3_V3 where
+  show = show.fromCl3_V3
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_V3 where
+  readPrec = toCl3_V3 <$> readPrec
+#endif
+
+instance Storable Cl3_V3 where
+  sizeOf _ = 3 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a1 <- peek (offset 0)
+    a2 <- peek (offset 1)
+    a3 <- peek (offset 2)
+    return $ Cl3_V3 a1 a2 a3
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_V3 a1 a2 a3) = do
+    poke (offset 0) a1
+    poke (offset 1) a2
+    poke (offset 2) a3
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_BV' a compact storable data type for BV.
+data Cl3_BV where
+  Cl3_BV :: !Double -> !Double -> !Double -> Cl3_BV
+
+-- | 'toCl3_BV' converts a Cl3 value constructed with BV to its compact form.
+toCl3_BV :: Cl3 -> Cl3_BV
+toCl3_BV (BV a23 a31 a12) = Cl3_BV a23 a31 a12
+toCl3_BV err = error $ "Please don't try and cast something that's not BV to Cl3_BV, Got: " ++ show err
+
+-- | 'fromCl3_BV' converts the compact Cl3_BV type back to a Cl3 type.
+fromCl3_BV :: Cl3_BV -> Cl3
+fromCl3_BV (Cl3_BV a23 a31 a12) = BV a23 a31 a12
+
+instance Show Cl3_BV where
+  show = show.fromCl3_BV
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_BV where
+  readPrec = toCl3_BV <$> readPrec
+#endif
+
+instance Storable Cl3_BV where
+  sizeOf _ = 3 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a23 <- peek (offset 0)
+    a31 <- peek (offset 1)
+    a12 <- peek (offset 2)
+    return $ Cl3_BV a23 a31 a12
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_BV a23 a31 a12) = do
+    poke (offset 0) a23
+    poke (offset 1) a31
+    poke (offset 2) a12
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_I' a compact storable data type for I.
+data Cl3_I where
+  Cl3_I :: !Double -> Cl3_I
+
+-- | 'toCl3_I' converts a Cl3 value constructed with I to its compact form.
+toCl3_I :: Cl3 -> Cl3_I
+toCl3_I (I a123) = Cl3_I a123
+toCl3_I err = error $ "Please don't try and cast something that's not R to Cl3_R, Got: " ++ show err
+
+-- | 'fromCl3_I' converts the compact Cl3_I type back to a Cl3 type.
+fromCl3_I :: Cl3_I -> Cl3
+fromCl3_I (Cl3_I a123) = I a123
+
+instance Show Cl3_I where
+  show = show.fromCl3_I
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_I where
+  readPrec = toCl3_I <$> readPrec
+#endif
+
+instance Storable Cl3_I where
+  sizeOf _ = sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a123 <- peek offset
+    return $ Cl3_I a123
+      where
+        offset = (castPtr ptr :: Ptr Double)
+
+  poke ptr (Cl3_I a123) = do
+    poke offset a123
+      where
+        offset = (castPtr ptr :: Ptr Double)
+
+
+-- | 'Cl3_PV' a compact storable data type for PV.
+data Cl3_PV where
+  Cl3_PV :: !Double -> !Double -> !Double -> !Double -> Cl3_PV
+
+-- | 'toCl3_PV' converts a Cl3 value constructed with PV to its compact form.
+toCl3_PV :: Cl3 -> Cl3_PV
+toCl3_PV (PV a0 a1 a2 a3) = Cl3_PV a0 a1 a2 a3
+toCl3_PV err = error $ "Please don't try and cast something that's not PV to Cl3_PV, Got: " ++ show err
+
+-- | 'fromCl3_PV' converts the compact Cl3_PV type back to a Cl3 type.
+fromCl3_PV :: Cl3_PV -> Cl3
+fromCl3_PV (Cl3_PV a0 a1 a2 a3) = PV a0 a1 a2 a3
+
+instance Show Cl3_PV where
+  show = show.fromCl3_PV
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_PV where
+  readPrec = toCl3_PV <$> readPrec
+#endif
+
+instance Storable Cl3_PV where
+  sizeOf _ = 4 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a0 <- peek (offset 0)
+    a1 <- peek (offset 1)
+    a2 <- peek (offset 2)
+    a3 <- peek (offset 4)
+    return $ Cl3_PV a0 a1 a2 a3
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_PV a0 a1 a2 a3) = do
+    poke (offset 0) a0
+    poke (offset 1) a1
+    poke (offset 2) a2
+    poke (offset 3) a3
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_H' a compact storable data type for H.
+data Cl3_H where
+  Cl3_H :: !Double -> !Double -> !Double -> !Double -> Cl3_H
+
+-- | 'toCl3_H' converts a Cl3 value constructed with H to its compact form.
+toCl3_H :: Cl3 -> Cl3_H
+toCl3_H (H a0 a23 a31 a12) = Cl3_H a0 a23 a31 a12
+toCl3_H err = error $ "Please don't try and cast something that's not H to Cl3_H, Got: " ++ show err
+
+-- | 'fromCl3_H' converts the compact Cl3_H type back to a Cl3 type.
+fromCl3_H :: Cl3_H -> Cl3
+fromCl3_H (Cl3_H a0 a23 a31 a12) = H a0 a23 a31 a12
+
+instance Show Cl3_H where
+  show = show.fromCl3_H
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_H where
+  readPrec = toCl3_H <$> readPrec
+#endif
+
+instance Storable Cl3_H where
+  sizeOf _ = 4 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a0 <- peek (offset 0)
+    a23 <- peek (offset 1)
+    a31 <- peek (offset 2)
+    a12 <- peek (offset 3)
+    return $ Cl3_H a0 a23 a31 a12
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_H a0 a23 a31 a12) = do
+    poke (offset 0) a0
+    poke (offset 1) a23
+    poke (offset 2) a31
+    poke (offset 3) a12
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_C' a compact storable data type for C.
+data Cl3_C where
+  Cl3_C :: !Double -> !Double -> Cl3_C
+
+-- | 'toCl3_C' converts a Cl3 value constructed with C to its compact form.
+toCl3_C :: Cl3 -> Cl3_C
+toCl3_C (C a0 a123) = Cl3_C a0 a123
+toCl3_C err = error $ "Please don't try and cast something that's not C to Cl3_C, Got: " ++ show err
+
+-- | 'fromCl3_C' converts the compact Cl3_C type back to a Cl3 type.
+fromCl3_C :: Cl3_C -> Cl3
+fromCl3_C (Cl3_C a0 a123) = C a0 a123
+
+instance Show Cl3_C where
+  show = show.fromCl3_C
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_C where
+  readPrec = toCl3_C <$> readPrec
+#endif
+
+instance Storable Cl3_C where
+  sizeOf _ = 2 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a0 <- peek (offset 0)
+    a123 <- peek (offset 1)
+    return $ Cl3_C a0 a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_C a0 a123) = do
+    poke (offset 0) a0
+    poke (offset 1) a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_BPV' a compact storable data type for BPV.
+data Cl3_BPV where
+  Cl3_BPV :: !Double -> !Double -> !Double -> !Double -> !Double -> !Double -> Cl3_BPV
+
+-- | 'toCl3_BPV' converts a Cl3 value constructed with BPV to its compact form.
+toCl3_BPV :: Cl3 -> Cl3_BPV
+toCl3_BPV (BPV a1 a2 a3 a23 a31 a12) = Cl3_BPV a1 a2 a3 a23 a31 a12
+toCl3_BPV err = error $ "Please don't try and cast something that's not BPV to Cl3_BPV, Got: " ++ show err
+
+-- | 'fromCl3_BPV' converts the compact Cl3_BPV type back to a Cl3 type.
+fromCl3_BPV :: Cl3_BPV -> Cl3
+fromCl3_BPV (Cl3_BPV a1 a2 a3 a23 a31 a12) = BPV a1 a2 a3 a23 a31 a12
+
+instance Show Cl3_BPV where
+  show = show.fromCl3_BPV
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_BPV where
+  readPrec = toCl3_BPV <$> readPrec
+#endif
+
+instance Storable Cl3_BPV where
+  sizeOf _ = 6 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a1 <- peek (offset 0)
+    a2 <- peek (offset 1)
+    a3 <- peek (offset 2)
+    a23 <- peek (offset 3)
+    a31 <- peek (offset 4)
+    a12 <- peek (offset 5)
+    return $ Cl3_BPV a1 a2 a3 a23 a31 a12
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_BPV a1 a2 a3 a23 a31 a12) = do
+    poke (offset 0) a1
+    poke (offset 1) a2
+    poke (offset 2) a3
+    poke (offset 3) a23
+    poke (offset 4) a31
+    poke (offset 5) a12
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_ODD' a compact storable data type for ODD.
+data Cl3_ODD where
+  Cl3_ODD :: !Double -> !Double -> !Double -> !Double -> Cl3_ODD
+
+-- | 'toCl3_ODD' converts a Cl3 value constructed with ODD to its compact form.
+toCl3_ODD :: Cl3 -> Cl3_ODD
+toCl3_ODD (ODD a1 a2 a3 a123) = Cl3_ODD a1 a2 a3 a123
+toCl3_ODD err = error $ "Please don't try and cast something that's not ODD to Cl3_ODD, Got: " ++ show err
+
+-- | 'fromCl3_ODD' converts the compact Cl3_ODD type back to a Cl3 type.
+fromCl3_ODD :: Cl3_ODD -> Cl3
+fromCl3_ODD (Cl3_ODD a1 a2 a3 a123) = ODD a1 a2 a3 a123
+
+instance Show Cl3_ODD where
+  show = show.fromCl3_ODD
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_ODD where
+  readPrec = toCl3_ODD <$> readPrec
+#endif
+
+instance Storable Cl3_ODD where
+  sizeOf _ = 4 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a1 <- peek (offset 0)
+    a2 <- peek (offset 1)
+    a3 <- peek (offset 2)
+    a123 <- peek (offset 3)
+    return $ Cl3_ODD a1 a2 a3 a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_ODD a1 a2 a3 a123) = do
+    poke (offset 0) a1
+    poke (offset 1) a2
+    poke (offset 2) a3
+    poke (offset 3) a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_TPV' a compact storable data type for TPV.
+data Cl3_TPV where
+  Cl3_TPV :: !Double -> !Double -> !Double -> !Double -> Cl3_TPV
+
+-- | 'toCl3_TPV' converts a Cl3 value constructed with TPV to its compact form.
+toCl3_TPV :: Cl3 -> Cl3_TPV
+toCl3_TPV (TPV a23 a31 a12 a123) = Cl3_TPV a23 a31 a12 a123
+toCl3_TPV err = error $ "Please don't try and cast something that's not TPV to Cl3_TPV, Got: " ++ show err
+
+-- | 'fromCl3_TPV' converts the compact Cl3_TPV type back to a Cl3 type.
+fromCl3_TPV :: Cl3_TPV -> Cl3
+fromCl3_TPV (Cl3_TPV a23 a31 a12 a123) = TPV a23 a31 a12 a123
+
+instance Show Cl3_TPV where
+  show = show.fromCl3_TPV
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_TPV where
+  readPrec = toCl3_TPV <$> readPrec
+#endif
+
+instance Storable Cl3_TPV where
+  sizeOf _ = 4 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a23 <- peek (offset 0)
+    a31 <- peek (offset 1)
+    a12 <- peek (offset 2)
+    a123 <- peek (offset 3)
+    return $ Cl3_TPV a23 a31 a12 a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_TPV a23 a31 a12 a123) = do
+    poke (offset 0) a23
+    poke (offset 1) a31
+    poke (offset 2) a12
+    poke (offset 3) a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+-- | 'Cl3_APS' a compact storable data type for APS.
+data Cl3_APS where
+  Cl3_APS :: !Double -> !Double -> !Double -> !Double -> !Double -> !Double -> !Double -> !Double -> Cl3_APS
+
+-- | 'toCl3_APS' converts a Cl3 value constructed with APS to its compact form.
+toCl3_APS :: Cl3 -> Cl3_APS
+toCl3_APS (APS a0 a1 a2 a3 a23 a31 a12 a123) = Cl3_APS a0 a1 a2 a3 a23 a31 a12 a123
+toCl3_APS err = error $ "Please don't try and cast something that's not APS to Cl3_APS, Got: " ++ show err
+
+-- | 'fromCl3_APS' converts the compact Cl3_APS type back to a Cl3 type.
+fromCl3_APS :: Cl3_APS -> Cl3
+fromCl3_APS (Cl3_APS a0 a1 a2 a3 a23 a31 a12 a123) = APS a0 a1 a2 a3 a23 a31 a12 a123
+
+instance Show Cl3_APS where
+  show = show.fromCl3_APS
+
+#ifndef O_NO_DERIVED
+instance Read Cl3_APS where
+  readPrec = toCl3_APS <$> readPrec
+#endif
+
+instance Storable Cl3_APS where
+  sizeOf _ = 8 * sizeOf (undefined :: Double)
+  alignment _ = sizeOf (undefined :: Double)
+  peek ptr = do
+    a0 <- peek (offset 0)
+    a1 <- peek (offset 1)
+    a2 <- peek (offset 2)
+    a3 <- peek (offset 3)
+    a23 <- peek (offset 4)
+    a31 <- peek (offset 5)
+    a12 <- peek (offset 6)
+    a123 <- peek (offset 7)
+    return $ Cl3_APS a0 a1 a2 a3 a23 a31 a12 a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+  poke ptr (Cl3_APS a0 a1 a2 a3 a23 a31 a12 a123) = do
+    poke (offset 0) a0
+    poke (offset 1) a1
+    poke (offset 2) a2
+    poke (offset 3) a3
+    poke (offset 4) a23
+    poke (offset 5) a31
+    poke (offset 6) a12
+    poke (offset 7) a123
+      where
+        offset i = (castPtr ptr :: Ptr Double) `plusPtr` (i*8)
+
+
+#endif
 
 
 
@@ -2476,8 +2939,8 @@ randUnitV3 :: RandomGen g => g -> (Cl3, g)
 randUnitV3 g =
   let (theta, g') = randomR (0,2*pi) g
       (u, g'') = randomR (-1,1) g'
-      simicircle = sqrt (1-u^2)
-  in (V3 (simicircle * cos theta) (simicircle * sin theta) u, g'')
+      semicircle = sqrt (1-u^2)
+  in (V3 (semicircle * cos theta) (semicircle * sin theta) u, g'')
 
 
 -- | 'randProjector' a projector with a random direction